Some updates to make this a LITTLE bit more modern.
No Hungarian notation.
PHP/CS fixes backported into code generator.

A very hackity-hack to allow types that are really arrays to be better
docblocked.

I've not seen an easy way to handle this automatically, so it's a real
nasty hack.

But the output passes PHPStan analysis without any additional rules.